requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.4.26
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
nodeenv==1.9.1
packaging==25.0
pip==24.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==4.2.0
pydantic==2.11.4
pydantic-core==2.33.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
pyyaml==6.0.2
ruff==0.11.10
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.1
uvicorn==0.34.2
virtualenv==20.31.2
